Move all users tests to tests dir
